fix: update @trpc/server imports to use dist
TRPC ships with its src ts files and these sneaky imports were accidentally targeting that ts, and not the dist files, making it so the whole trpc lib was getting checked by the consumers tsc build! In particular, this surfaced for me because my project has verbatimModuleSyntax set to true, and trpc does not lean on that option.